myportfolio/requirements.txt
asgiref==3.2.10
astroid==2.4.2
colorama==0.4.3
Django==3.1
django-environ==0.4.5
django-ordered-model==3.4.1
dj-database-url==0.5.0
gunicorn==20.0.4
isort==5.4.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysqlclient==2.0.1
psycopg2==2.8.5
pylint==2.6.0
python-dotenv==0.14.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
whitenoise==5.2.0
wrapt==1.12.1